src/auth/requirements.txt
python-dotenv
requests
fastapi
uvicorn
redis
